srcs/requirements/auth/django/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
django-cors-headers==4.4.0
